Move `OpenAuditLogFileUntilSuccess` to new `helpers` package
This function is relevant also outside of the `ginaudit` package, so
let's make it available through a new package.
