+++ ginaudit/mdw_test.go
@@ -29,7 +29,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
-	"path/filepath"
 	"regexp"
 	"strings"
 	"sync"
@@ -317,68 +316,6 @@ func TestParallelCallsToMiddleware(t *testing.T) {
 
 		require.Regexp(t, regexp.MustCompile(metricToCompare), str)
 	}
-}
-
-func TestOpenAuditLogFileUntilSuccess(t *testing.T) {
-	t.Parallel()
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-
-	tmpdir := t.TempDir()
-	tmpfile := filepath.Join(tmpdir, "audit.log")
-
-	go func() {
-		defer wg.Done()
-		time.Sleep(time.Second)
-		fd, err := os.OpenFile(tmpfile, os.O_RDONLY|os.O_CREATE, 0o600)
-		require.NoError(t, err)
-		err = fd.Close()
-		require.NoError(t, err)
-	}()
-
-	fd, err := ginaudit.OpenAuditLogFileUntilSuccess(tmpfile)
-	require.NoError(t, err)
-	require.NotNil(t, fd)
-
-	err = fd.Close()
-	require.NoError(t, err)
-
-	// We wait so we don't leak file descriptors
-	wg.Wait()
-
-	err = os.Remove(tmpfile)
-	require.NoError(t, err)
-}
-
-func TestOpenAuditLogFileError(t *testing.T) {
-	t.Parallel()
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-
-	tmpdir := t.TempDir()
-	tmpfile := filepath.Join(tmpdir, "audit.log")
-
-	go func() {
-		defer wg.Done()
-		time.Sleep(time.Second)
-		// This file is read only
-		fd, err := os.OpenFile(tmpfile, os.O_RDONLY|os.O_CREATE, 0o500)
-		require.NoError(t, err)
-		err = fd.Close()
-		require.NoError(t, err)
-	}()
-
-	fd, err := ginaudit.OpenAuditLogFileUntilSuccess(tmpfile)
-	require.Error(t, err)
-	require.Nil(t, fd)
-
-	// We wait so we don't leak file descriptors
-	wg.Wait()
-
-	err = os.Remove(tmpfile)
-	require.NoError(t, err)
 }
 
 func TestCantRegisterMultipleTimesToSamePrometheus(t *testing.T) {
